packaging/fedora: Add Amazon Linux 2 support
Amazon Linux 2 is a derivative of Red Hat Enterprise Linux 7, using
a combination of CentOS 7 sources, backports from Fedora, and custom
packages produced by the Amazon Linux team to provide an updated stack
on a stable base platform.

As the Fedora packaging is already most of the way there for supporting
RHEL/CentOS 7, we can easily enable Amazon Linux 2, as well.

Note that the packaging assumes that EPEL for EL7 is activated on Amazon
Linux 2, as we require packages from EPEL for installation.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>